# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: KlassenPilot API
  version: 0.1.0
paths:
...
  /api/classes/{classId}/course/network:
    get:
      summary: Get an adopted class course network
      parameters:
      - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Adopted network or null when none has been adopted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/classes/{classId}/course/network/sources/{sourceId}/sections/{sectionId}:
    get:
      summary: Read exact class-authorized curriculum evidence for a course network
      parameters:
      - $ref: '#/components/parameters/ClassId'
      - $ref: '#/components/parameters/SourceId'
      - $ref: '#/components/parameters/SectionId'
      responses:
        '200':
          description: Genuine source section and provenance authorized for the class route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkSourceSectionResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/classes/{classId}/course/network/drafts:
    post:
      summary: Open the matching reviewed course-network seed as a structured draft
      parameters:
      - $ref: '#/components/parameters/ClassId'
      responses:
        '201':
          description: Structured proposed seed draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '409':
          description: A network was already adopted for this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/classes/{classId}/course/network/drafts/{draftId}:
    get:
      summary: Get a structured course-network seed draft
      parameters:
      - $ref: '#/components/parameters/ClassId'
      - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: The persisted draft artifact and its latest review snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/classes/{classId}/course/network/drafts/{draftId}/review:
    post:
      summary: Validate and no-tools review the exact structured seed draft
      parameters:
      - $ref: '#/components/parameters/ClassId'
      - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Completed deterministic and no-tools review bound to this artifact revision/hash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/classes/{classId}/course/network/drafts/{draftId}/adopt:
    post:
      summary: Teacher-approved adoption of the exact reviewed seed revision
      parameters:
      - $ref: '#/components/parameters/ClassId'
      - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoptCourseNetworkDraftRequest'
      responses:
        '200':
          description: Adopted canonical course network
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkAdoptionResponse'
        '409':
          description: Stale review, non-accepted review, or duplicate adoption
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/classes/{class_id}/course/network/drafts/{draft_id}/revise:
    post:
      tags:
      - course-network
      summary: Revise Seed
      operationId: revise_seed_api_classes__class_id__course_network_drafts__draft_id__revise_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseSeedRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Draft is terminal, adoption is in progress, or review/artifact snapshot is stale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Unusable or timed-out model response; retry without changing memory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/network:
    get:
      tags:
      - course-network
      summary: Get Course Network
      operationId: get_course_network_api_classes__class_id__course_network_get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/network/sources/{source_id}/sections/{section_id}:
    get:
      tags:
      - course-network
      summary: Get Course Network Source Section
      operationId: get_course_network_source_section_api_classes__class_id__course_network_sources__source_id__sections__section_id__get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          title: Source Id
      - name: section_id
        in: path
        required: true
        schema:
          type: string
          title: Section Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkSourceSectionResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/network/drafts:
    post:
      tags:
      - course-network
      summary: Open Course Network Seed Draft
      operationId: open_course_network_seed_draft_api_classes__class_id__course_network_drafts_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: A network was already adopted for this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/network/drafts/{draft_id}:
    get:
      tags:
      - course-network
      summary: Get Course Network Draft
      operationId: get_course_network_draft_api_classes__class_id__course_network_drafts__draft_id__get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/network/drafts/{draft_id}/review:
    post:
      tags:
      - course-network
      summary: Review Course Network Seed
      operationId: review_course_network_seed_api_classes__class_id__course_network_drafts__draft_id__review_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkDraftResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Draft is terminal, adoption is in progress, or review/artifact snapshot is stale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/network/drafts/{draft_id}/adopt:
    post:
      tags:
      - course-network
      summary: Adopt Course Network Seed
      operationId: adopt_course_network_seed_api_classes__class_id__course_network_drafts__draft_id__adopt_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoptCourseNetworkDraftRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNetworkAdoptionResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, non-accepted review, or duplicate adoption
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/classes/{class_id}/course/network/nodes/{node_id}/lessons:
    get:
      tags:
      - course-network
      summary: Get Course Node Lessons
      operationId: get_course_node_lessons_api_classes__class_id__course_network_nodes__node_id__lessons_get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseNodeLessonsResponse'
        '404':
          description: Class, course-network draft, or authorized evidence was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request or structured draft validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    ClassId:
      name: classId
      in: path
      required: true
      schema:
        type: string
      description: Stable class identifier.
    DraftId:
      name: draftId
      in: path
      required: true
      schema:
        type: string
      description: Durable structured course-network workflow draft identifier.
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
      description: Stable trusted-source identifier authorized for the class route.
    SectionId:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
      description: Exact trusted-source section identifier.
  responses:
    NotFoundError:
      description: Class, course-network draft, or authorized evidence was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ConflictError:
      description: Draft is terminal, adoption is in progress, or review/artifact snapshot is stale.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: Request or structured draft validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorBody:
      properties:
        type:
          type: string
          title: Type
        message:
          type: string
          title: Message
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
      type: object
      required:
      - type
      - message
      title: ErrorBody
    ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - error
      title: ErrorEnvelope
    CourseNetworkResponse:
      properties:
        class_id:
          type: string
          title: Class Id
        network:
          anyOf:
          - $ref: '#/components/schemas/CourseNetworkDocument'
          - type: 'null'
      type: object
      required:
      - class_id
      title: CourseNetworkResponse
...
    CourseNetworkSourceProvenance:
      properties:
        authority:
          type: string
          title: Authority
        jurisdiction:
          type: string
          title: Jurisdiction
        canonical_url:
          type: string
          title: Canonical Url
        retrieved_at:
          type: string
          title: Retrieved At
        version_label:
          type: string
          title: Version Label
        content_hash:
          type: string
          title: Content Hash
      type: object
      required:
      - authority
      - jurisdiction
      - canonical_url
      - retrieved_at
      - version_label
      - content_hash
      title: CourseNetworkSourceProvenance
    CourseNetworkSourceSectionResponse:
      properties:
        source_id:
          type: string
          title: Source Id
        source_title:
          type: string
          title: Source Title
        section_id:
          type: string
          title: Section Id
        section_title:
          type: string
          title: Section Title
        content:
          type: string
          title: Content
        provenance:
          $ref: '#/components/schemas/CourseNetworkSourceProvenance'
      type: object
      required:
      - source_id
      - source_title
      - section_id
      - section_title
      - content
      - provenance
      title: CourseNetworkSourceSectionResponse
...
    CanvasPosition:
      properties:
        x:
          type: number
          title: X
        y:
          type: number
          title: Y
      type: object
      required:
      - x
      - y
      title: CanvasPosition
    CourseNetworkDocument:
      properties:
        schema_version:
          type: integer
          const: 1
          title: Schema Version
          default: 1
        class_id:
          type: string
          title: Class Id
        route:
          $ref: '#/components/schemas/CurriculumRouteRef'
        revision:
          type: integer
          minimum: 1.0
          title: Revision
          default: 1
        nodes:
          items:
            $ref: '#/components/schemas/LearningBlock'
          type: array
          title: Nodes
        edges:
          items:
            $ref: '#/components/schemas/NetworkEdge'
          type: array
          title: Edges
        material_mappings:
          items:
            $ref: '#/components/schemas/MaterialMapping'
          type: array
          title: Material Mappings
        positions:
          additionalProperties:
            $ref: '#/components/schemas/CanvasPosition'
          type: object
          title: Positions
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - class_id
      - route
      - updated_at
      title: CourseNetworkDocument
      description: 'Canonical course-network data for one class.


        ``for_draft_seed`` is the sole construction path that permits proposed

        nodes. Ordinary construction represents a canonical document and rejects

        them before any durable write can occur.'
...
    CourseNetworkDraftResponse:
      properties:
        draft_id:
          type: string
          title: Draft Id
        class_id:
          type: string
          title: Class Id
        status:
          type: string
          title: Status
        artifact_markdown:
          type: string
          title: Artifact Markdown
        artifact_revision:
          type: integer
          title: Artifact Revision
        artifact_hash:
          type: string
          title: Artifact Hash
        backend_session_id:
          type: string
          title: Backend Session Id
        network:
          $ref: '#/components/schemas/CourseNetworkDraftDocument'
        review:
          anyOf:
          - $ref: '#/components/schemas/CourseNetworkReviewResult'
          - type: 'null'
      type: object
      required:
      - draft_id
      - class_id
      - status
      - artifact_markdown
      - artifact_revision
      - artifact_hash
      - backend_session_id
      - network
      title: CourseNetworkDraftResponse
    AdoptCourseNetworkDraftRequest:
      properties:
        expected_revision:
          type: integer
          minimum: 0.0
          title: Expected Revision
        expected_hash:
          type: string
          minLength: 1
          title: Expected Hash
      type: object
      required:
      - expected_revision
      - expected_hash
      title: AdoptCourseNetworkDraftRequest
    CourseNetworkAdoptionResponse:
      properties:
        class_id:
          type: string
          title: Class Id
        draft_id:
          type: string
          title: Draft Id
        log_entry_id:
          type: string
          title: Log Entry Id
        network:
          $ref: '#/components/schemas/CourseNetworkDocument'
      type: object
      required:
      - class_id
      - draft_id
      - log_entry_id
      - network
      title: CourseNetworkAdoptionResponse
...
    ReviseSeedRequest:
      properties:
        teacher_request:
          type: string
          maxLength: 6000
          title: Teacher Request
          default: Resolve the current review findings with minimal changes.
        expected_revision:
          type: integer
          title: Expected Revision
        expected_hash:
          type: string
          title: Expected Hash
      type: object
      required:
      - expected_revision
      - expected_hash
      title: ReviseSeedRequest
...
    CourseNodeLessonsResponse:
      properties:
        class_id:
          type: string
          title: Class Id
        node_id:
          type: string
          title: Node Id
        associations:
          items:
            $ref: '#/components/schemas/CourseLessonAssociation'
          type: array
          title: Associations
      type: object
      required:
      - class_id
      - node_id
      - associations
      title: CourseNodeLessonsResponse
...